# crates/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sentrix Chain Indexer API",
    "version": "0.1.0",
    "description": "Read-only REST + GraphQL surface over the Sentrix Chain indexer DB. Mirrors the TS Fastify port byte-for-byte on response shape; production-grade layers (cache, rate-limit, auth, metrics, readyz) added on top.\n\nWire conventions:\n- All bigint fields (block heights, timestamps, gas, value, fee) render as decimal strings to survive JSON's f64 precision limit.\n- `from_addr` / `to_addr` renamed to `from` / `to` in the wire shape.\n- Log topics flattened to a `topics` array with null entries dropped.\n- Address inputs are lowercased server-side.",
    "license": { "name": "BUSL-1.1" }
  },
  "servers": [
    { "url": "https://api.sentrixchain.com", "description": "Mainnet" },
    { "url": "https://testnet-api.sentrixchain.com", "description": "Testnet" },
    { "url": "http://localhost:8080", "description": "Local dev" }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Optional. Set INDEXER_API_BEARER_TOKEN on the server to require `Authorization: Bearer <token>` on every route except /health. Default deploys leave this off and gate auth at the Caddy edge."
      }
    },
    "parameters": {
      "limit": {
        "name": "limit",
        "in": "query",
        "description": "Page size, clamped 1..=100, default 25.",
        "required": false,
        "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 25 }
      },
      "before": {
        "name": "before",
        "in": "query",
        "description": "Block-height upper bound (inclusive). Use the `next_cursor` value from the previous page response.",
        "required": false,
        "schema": { "type": "string" }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error", "code"],
        "properties": {
          "error": { "type": "string", "description": "Human-readable message." },
          "code":  { "type": "string", "enum": ["invalid_query", "not_found", "internal", "unauthorized"] },
          "request_id": { "type": "string", "description": "Echo of x-request-id header for log correlation." }
        }
      },
      "Block": {
        "type": "object",
        "required": ["height","hash","parent_hash","timestamp","validator","gas_used","gas_limit","tx_count","round"],
        "properties": {
          "height":       { "type": "string", "description": "Decimal-string bigint" },
          "hash":         { "type": "string", "pattern": "^0x[0-9a-f]{64}$" },
          "parent_hash":  { "type": "string", "pattern": "^0x[0-9a-f]{64}$" },
          "timestamp":    { "type": "string", "description": "Unix seconds, decimal string" },
          "validator":    { "type": "string", "pattern": "^0x[0-9a-f]{40}$" },
          "gas_used":     { "type": "string" },
          "gas_limit":    { "type": "string" },
          "base_fee":     { "type": ["string", "null"] },
          "tx_count":     { "type": "integer" },
          "state_root":   { "type": ["string", "null"] },
          "round":        { "type": "integer" }
        }
      },
...
      "BlocksList": {
        "type": "object",
        "required": ["blocks"],
        "properties": {
          "blocks":      { "type": "array", "items": { "$ref": "#/components/schemas/Block" } },
          "next_cursor": { "type": ["string", "null"], "description": "Pass back as ?before= to fetch the next page. Null when the page reached block 0." }
        }
      }
    }
  },
  "paths": {
...
    "/blocks": {
      "get": {
        "tags": ["blocks"], "summary": "Latest blocks (paginated)",
        "parameters": [
          { "$ref": "#/components/parameters/limit" },
          { "$ref": "#/components/parameters/before" }
        ],
        "responses": {
          "200": { "description": "ok", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BlocksList" } } } },
          "400": { "description": "invalid `before`", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/blocks/{height}": {
      "get": {
        "tags": ["blocks"], "summary": "Block detail with nested transactions",
        "parameters": [{ "name": "height", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "ok" },
          "400": { "description": "invalid height", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "404": { "description": "no block at this height", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  }
}